double dmr test with threshold

+++ src/cuda/gemm_tensorcores/gemm_optimized/main.cpp
@@ -16,8 +16,8 @@
 
 #define THRESHOLD 1.0E-6  
 
-typedef float real_t;
-typedef float half_real_t;
+typedef double real_t;
+typedef double half_real_t;
 
 void gemm_host(std::vector<real_t>& a, std::vector<real_t>& b,
 		std::vector<real_t>& c, real_t alpha, real_t beta, int m, int n,
